Use new dtype hierarchy from NEP 41

+++ tests/pass/simple_py3.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-array = np.array([1, 2])
+array = np.array([1, 2], dtype=np.int64)
 
 # The @ operator is not in python 2
 array @ array
